Release 1.0.31

+++ pkg/ast/pipesearch/searchHandler.go
@@ -643,7 +643,11 @@ func RunQueryForNewPipeline(conn *websocket.Conn, qid uint64, root *structs.ASTN
 			continue
 		}
 
+		rQuery.SetLatestQueryState(queryStateData.StateName)
+
 		switch queryStateData.StateName {
+		case query.WAITING:
+			// do nothing
 		case query.READY:
 			switch queryStateData.ChannelIndex {
 			case multiplexer.MainIndex:
@@ -669,7 +673,9 @@ func RunQueryForNewPipeline(conn *websocket.Conn, qid uint64, root *structs.ASTN
 			}
 		case query.COMPLETE:
 			defer query.DeleteQuery(qid)
-			defer query.DeleteQuery(timechartQid)
+			if runTimechartQuery {
+				defer query.DeleteQuery(timechartQid)
+			}
 
 			if isAsync {
 				wErr := conn.WriteJSON(queryStateData.CompleteWSResp)
@@ -704,7 +710,9 @@ func RunQueryForNewPipeline(conn *websocket.Conn, qid uint64, root *structs.ASTN
 			}
 
 			defer query.DeleteQuery(qid)
-			defer query.DeleteQuery(timechartQid)
+			if runTimechartQuery {
+				defer query.DeleteQuery(timechartQid)
+			}
 			if isAsync {
 				wErr := conn.WriteJSON(createErrorResponse(queryStateData.Error.Error()))
 				if wErr != nil {
@@ -740,18 +748,24 @@ func RunQueryForNewPipeline(conn *websocket.Conn, qid uint64, root *structs.ASTN
 			defer query.DeleteQuery(timechartQid)
 			if isAsync {
 				processTimeoutUpdate(conn, qid)
-				processTimeoutUpdate(conn, timechartQid)
+				if runTimechartQuery {
+					processTimeoutUpdate(conn, timechartQid)
+				}
 			} else {
 				return nil, false, root.TimeRange, fmt.Errorf("qid=%v, RunQueryForNewPipeline: query timed out", qid)
 			}
 		case query.CANCELLED:
 			log.Infof("qid=%v, RunQueryForNewPipeline: query cancelled", qid)
 			defer query.DeleteQuery(qid)
-			defer query.DeleteQuery(timechartQid)
+			if runTimechartQuery {
+				defer query.DeleteQuery(timechartQid)
+			}
 
 			if isAsync {
 				processCancelQuery(conn, qid)
-				processCancelQuery(conn, timechartQid)
+				if runTimechartQuery {
+					processCancelQuery(conn, timechartQid)
+				}
 			}
 			return nil, false, root.TimeRange, nil
 		}
